+++ src/readi2cblock.cc
@@ -66,9 +66,9 @@ NAN_METHOD(ReadI2cBlockAsync) {
       "(int fd, int cmd, int length, Buffer buffer, function cb)"));
   }
 
-  int fd = info[0]->Int32Value();
-  __u8 cmd = info[1]->Int32Value();
-  __u32 length = info[2]->Uint32Value();
+  int fd = Nan::To<int32_t>(info[0]).FromJust();
+  __u8 cmd = Nan::To<int32_t>(info[1]).FromJust();
+  __u32 length = Nan::To<uint32_t>(info[2]).FromJust();
   v8::Local<v8::Object> bufferHandle = info[3].As<v8::Object>();
   Nan::Callback *callback = new Nan::Callback(info[4].As<v8::Function>());
 
@@ -106,9 +106,9 @@ NAN_METHOD(ReadI2cBlockSync) {
       "(int fd, int cmd, int length, Buffer buffer)"));
   }
 
-  int fd = info[0]->Int32Value();
-  __u8 cmd = info[1]->Int32Value();
-  __u32 length = info[2]->Uint32Value();
+  int fd = Nan::To<int32_t>(info[0]).FromJust();
+  __u8 cmd = Nan::To<int32_t>(info[1]).FromJust();
+  __u32 length = Nan::To<uint32_t>(info[2]).FromJust();
   v8::Local<v8::Object> bufferHandle = info[3].As<v8::Object>();
 
   __u8* bufferData = (__u8*) node::Buffer::Data(bufferHandle);